+++ src/PropertyDescriber/FloatPropertyDescriber.php
@@ -21,6 +21,14 @@ class FloatPropertyDescriber implements PropertyDescriberInterface
      */
     public function describe(array $types, OA\Schema $property, ?array $groups = null, ?OA\Schema $schema = null, array $context = [])
     {
+        if (null !== $groups) {
+            trigger_deprecation(
+                'nelmio/api-doc-bundle',
+                '4.17.0',
+                'Using the $groups property is deprecated and will be removed in a future version. Pass groups via $context[\'groups\']',
+            );
+        }
+
         $property->type = 'number';
         $property->format = 'float';
     }