+++ src/PropertyDescriber/RequiredPropertyDescriber.php
@@ -26,6 +26,14 @@ final class RequiredPropertyDescriber implements PropertyDescriberInterface, Pro
      */
     public function describe(array $types, OA\Schema $property, ?array $groups = null, ?OA\Schema $schema = null, array $context = [])
     {
+        if (null !== $groups) {
+            trigger_deprecation(
+                'nelmio/api-doc-bundle',
+                '4.17.0',
+                'Using the $groups property is deprecated and will be removed in a future version. Pass groups via $context[\'groups\']',
+            );
+        }
+
         $this->propertyDescriber->describe($types, $property, $groups, $schema, $context);
 
         if (!$property instanceof OA\Property) {