+++ CHANGELOG.md
@@ -135,6 +135,10 @@ doc-api:
 * Added Redocly as an alternative to Swagger UI. https://github.com/Redocly/redoc.
 * Added support for describing dictionary types in OpenAPI 3.0.
 
+4.17.0
+-----
+* Passing groups to `PropertyDescriberInterface::describe()` via the `$groups` property is deprecated, the property will get removed in a future version. Pass groups via `$context['groups']` instead.
+
 4.0.0
 -----
 * Added support of OpenAPI 3.0. The internals were completely reworked and this version introduces BC breaks.
@@ -162,7 +166,7 @@ doc-api:
 
 * Add a documentation form extension. Use the ``documentation`` option to define how a form field is documented.
 * Allow references to config definitions in controllers.
-* Using `@Model` implicitely in `@SWG\Schema`, `@SWG\Items` and `@SWG\Property` is deprecated. Use `ref=@Model()` instead.
+* Using `@Model` implicitly in `@SWG\Schema`, `@SWG\Items` and `@SWG\Property` is deprecated. Use `ref=@Model()` instead.
 
   Before:
   ```php